gtdf

+++ CaiJinTongApp/Interface/CommentListInterface.h
@@ -15,7 +15,7 @@
 
 @property (nonatomic, assign) id<CommentListInterfaceDelegate>delegate;
 
--(void)getGradeInterfaceDelegateWithUserId:(NSString *)userId andSectionId:(NSString *)sectionId andPageIndex:(int)pageIndex;
+-(void)getGradeInterfaceDelegateWithUserId:(NSString *)userId andSectionId:(NSString *)sectionId andPageIndex:(NSInteger)pageIndex;
 @end
 
 @protocol CommentListInterfaceDelegate <NSObject>
